requirements.txt
Flask==2.3.3
numpy==1.24.3
tflite_runtime
pillow==10.0.1
gunicorn==21.2.0